Felület

+++ MicroSim/MicroSim/Entities/Deathprobability.cs
@@ -9,7 +9,7 @@ namespace MicroSim.Entities
     public class Deathprobability
     {
         public Gender Gender { get; set; }
-        public int Age { get; set; }
+        public byte Age { get; set; }
         public double P { get; set; }
     }
 }
